Correcting Client Compile Errors & Warnings
Not quite done -- need to actually implement the 'metastats' array in the client, and test that it pulls from the server correctly.

+++ Source/Client/common.h
@@ -219,7 +219,7 @@ struct cplayer {
 
 	int points;
 	int points_tot;
-	int kindred;
+	unsigned long kindred;
 
 	// possessions
 	int gold;
